downloads

+++ ovos_skills_manager/skill_entry.py
@@ -137,8 +137,14 @@ class SkillEntry:
 
     @property
     def download_url(self):
+        """ provided in .json file """
         return self.json.get("download_url") or \
                download_url_from_github_url(self.url, self.branch)
+
+    @property
+    def default_download_url(self):
+        """ generated from github url directly"""
+        return download_url_from_github_url(self.url, self.branch)
 
     @property
     def requirements(self):
@@ -288,7 +294,8 @@ class SkillEntry:
                 with open(icon_file, "wb") as f:
                     f.write(content)
             else:
-                shutil.copyfile(self.skill_icon, icon_file)
+                if isfile(self.skill_icon):
+                    shutil.copyfile(self.skill_icon, icon_file)
 
             # copy .desktop file
             desktop_file = join(desktop_dir, base_name + ".desktop")
